move logicFromInsight to another file to avoid recursive imports of funnelLogic and insightLogic
